Fix type unification bug which caused incorrect type inference
For example, '[apply] apply' would incorrectly have its type inferred
as 'a->b', not 'a*(a->b) -> b'.

This bug existed because, when unifying complex forms (function or
product types), the substitutions derived in the first pair of types
weren't applied to the types in the second pair before attempting to unify
the second pair.